requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0
folium>=0.12.0
geopandas>=0.11.0
requests>=2.28.0
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.7.0
python-dotenv>=1.0.0
us>=3.0.0